feat: 禁用模组
